Suffix Tree. Ukkonen's Algorithm
This algorithm builds a suffix tree for a given string s of length n in O(nlog(k))) time, where k is the size of the alphabet (if k is considered to be a constant, the asymptotic behavior is linear).

The input to the algorithm are the string s and its length n, which are passed as global variables.

The main function build_tree builds a suffix tree. It is stored as an array of structures node, where node[0] is the root of the tree.

In order to simplify the code, the edges are stored in the same structures: for each vertex its structure node stores the information about the edge between it and its parent. Overall each node stores the following information:

(l, r) - left and right boundaries of the substring s[l..r-1] which correspond to the edge to this node,
par - the parent node,
link - the suffix link,
next - the list of edges going out from this node.